+++ Android/app/src/main/java/com/nice295/memo/NewRecordActivity.java
@@ -1,17 +1,32 @@
 package com.nice295.memo;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.View;
+import android.widget.Button;
 
-/**
- * Created by kyuholee on 2016. 7. 30..
- */
 public class NewRecordActivity extends AppCompatActivity {
-    private static final String TAG = "NewRecordActivity";
-
+    private  static Button button;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_new_record);
+        OnClickButtonListener();
+
+
+    }
+
+    public void OnClickButtonListener(){
+        button = (Button)findViewById(R.id.button_first);
+        button.setOnClickListener(
+                new View.OnClickListener(){
+                    @Override
+                    public void onClick(View v){
+                        Intent intent =new Intent("com.example.android.record_160806.Record_second");
+                        startActivity(intent);
+                    }
+                }
+        );
     }
 }
