+++ Android/app/src/main/res/layout/activity_new_record_2.xml
