+++ Android/app/src/main/res/layout/record_complete.xml
