+++ Android/app/src/main/res/layout/record_message.xml
