--- Record_160808/.idea/.name
@@ -1,1 +1,0 @@
-Record_160808
