--- Record_160808/settings.gradle
@@ -1,1 +1,0 @@
-include ':app'
